refactor: "테마가 존재하지 않을 경우 예약 불가능" 로직 서비스 계층으로 이동
